generate include/ode/precision.h to set the precision instead of command-line defines; general code cleanup to get rid of GCC warnings

+++ ode/src/threading_impl_templates.h
@@ -43,6 +43,8 @@
 
 #include <new>
 
+#include <inttypes.h>
+
 
 #define dMAKE_JOBINSTANCE_RELEASEE(job_instance) ((dCallReleaseeID)(job_instance))
 #define dMAKE_RELEASEE_JOBINSTANCE(releasee) ((dxThreadedJobInfo *)(releasee))
@@ -72,7 +74,7 @@ private:
     union
     {
         dmutexindex_t     m_mutex_count;
-        unsigned long long m_reserved_for_allignment;
+        uint64_t m_reserved_for_allignment;
 
     } m_un;
